split transport tests
