+++ src/layers/6_client/client.test.ts
@@ -3,10 +3,6 @@ import { createResponse, test } from '../../../tests/_/helpers.js'
 import { Graffle as Graffle2 } from '../../../tests/_/schema/generated/__.js'
 import { schema } from '../../../tests/_/schema/schema.js'
 import { Graffle } from '../../entrypoints/main.js'
-import { ACCEPT_REC, CONTENT_TYPE_REC } from '../../lib/graphqlHTTP.js'
-import { CONTENT_TYPE_GQL, CONTENT_TYPE_JSON } from '../../lib/http.js'
-import { Transport } from '../5_core/types.js'
-import type { RequestInput } from './Settings/inputIncrementable/request.js'
 
 const endpoint = new URL(`https://foo.io/api/graphql`)
 
@@ -27,107 +23,6 @@ describe(`without schemaIndex only raw is available`, () => {
   test(`available methods`, () => {
     expect(graffle.raw).toBeTypeOf(`function`) // eslint-disable-line
     expect(graffle.rawOrThrow).toBeTypeOf(`function`) // eslint-disable-line
-  })
-})
-
-describe(`transport`, () => {
-  describe(`http`, () => {
-    test(`anyware hooks are typed to http transport`, () => {
-      Graffle.create({ schema: endpoint }).use(async ({ encode }) => {
-        expectTypeOf(encode.input.transport).toEqualTypeOf(Transport.http)
-        const { pack } = await encode()
-        expectTypeOf(pack.input.transport).toEqualTypeOf(Transport.http)
-        const { exchange } = await pack()
-        expectTypeOf(exchange.input.transport).toEqualTypeOf(Transport.http)
-        expectTypeOf(exchange.input.request).toEqualTypeOf<RequestInput>()
-        const { unpack } = await exchange()
-        expectTypeOf(unpack.input.transport).toEqualTypeOf(Transport.http)
-        expectTypeOf(unpack.input.response).toEqualTypeOf<Response>()
-        const { decode } = await unpack()
-        expectTypeOf(decode.input.transport).toEqualTypeOf(Transport.http)
-        expectTypeOf(decode.input.response).toEqualTypeOf<Response>()
-        const result = await decode()
-        if (!(result instanceof Error)) {
-          expectTypeOf(result.response).toEqualTypeOf<Response>()
-        }
-        return result
-      })
-    })
-    test(`can set headers in constructor`, async ({ fetch }) => {
-      fetch.mockImplementationOnce(() => Promise.resolve(createResponse({ data: { id: `abc` } })))
-      const graffle = Graffle.create({ schema: endpoint, request: { headers: { 'x-foo': `bar` } } })
-      await graffle.rawString({ document: `query { id }` })
-      const request = fetch.mock.calls[0]?.[0]
-      expect(request?.headers.get(`x-foo`)).toEqual(`bar`)
-    })
-    test(`sends spec compliant request`, async ({ fetch, graffle }) => {
-      fetch.mockImplementationOnce(() => Promise.resolve(createResponse({ data: { greetings: `Hello World` } })))
-      await graffle.rawString({ document: `query { greetings }` })
-      const request = fetch.mock.calls[0]?.[0]
-      expect(request?.headers.get(`content-type`)).toEqual(CONTENT_TYPE_REC)
-      expect(request?.headers.get(`accept`)).toEqual(ACCEPT_REC)
-    })
-    describe(`signal`, () => {
-      // JSDom and Node result in different errors. JSDom is a plain Error type. Presumably an artifact of JSDom and now in actual browsers.
-      const abortErrorMessagePattern = /This operation was aborted|AbortError: The operation was aborted/
-      test(`AbortController at instance level works`, async () => {
-        const abortController = new AbortController()
-        const graffle = Graffle.create({
-          schema: endpoint,
-          request: { signal: abortController.signal },
-        })
-        const resultPromise = graffle.rawString({ document: `query { id }` })
-        abortController.abort()
-        const { caughtError } = await resultPromise.catch((caughtError: unknown) => ({ caughtError })) as any as {
-          caughtError: Error
-        }
-        expect(caughtError.message).toMatch(abortErrorMessagePattern)
-      })
-      test(`AbortController at method level works`, async () => {
-        const abortController = new AbortController()
-        const graffle = Graffle.create({
-          schema: endpoint,
-        }).with({ request: { signal: abortController.signal } })
-        const resultPromise = graffle.rawString({ document: `query { id }` })
-        abortController.abort()
-        const { caughtError } = await resultPromise.catch((caughtError: unknown) => ({ caughtError })) as any as {
-          caughtError: Error
-        }
-        expect(caughtError.message).toMatch(abortErrorMessagePattern)
-      })
-    })
-  })
-  describe(`memory`, () => {
-    test(`anyware hooks are typed to memory transport`, () => {
-      Graffle.create({ schema }).use(async ({ encode }) => {
-        expectTypeOf(encode.input.transport).toEqualTypeOf(Transport.memory)
-        const { pack } = await encode()
-        expectTypeOf(pack.input.transport).toEqualTypeOf(Transport.memory)
-        const { exchange } = await pack()
-        expectTypeOf(exchange.input.transport).toEqualTypeOf(Transport.memory)
-        // @ts-expect-error any
-        exchange.input.request
-        const { unpack } = await exchange()
-        expectTypeOf(unpack.input.transport).toEqualTypeOf(Transport.memory)
-        // @ts-expect-error any
-        unpack.input.response
-        const { decode } = await unpack()
-        expectTypeOf(decode.input.transport).toEqualTypeOf(Transport.memory)
-        // @ts-expect-error any
-        decode.input.response
-        const result = await decode()
-        if (!(result instanceof Error)) {
-          // @ts-expect-error any
-          result.response
-        }
-        return result
-      })
-    })
-    test(`cannot set headers in constructor`, () => {
-      // todo: This error is poor for the user. It refers to schema not being a URL. The better message would be that headers is not allowed with memory transport.
-      // @ts-expect-error headers not allowed with GraphQL schema
-      Graffle.create({ schema, request: { headers: { 'x-foo': `bar` } } })
-    })
   })
 })
